deleted double servers and refactoring code
