+++ test/handlers/get-issues.test.js
@@ -1,26 +1,29 @@
 var test = require("tape");
-var server = require("../../issuesAPI/server.js");
+var config = require("../../config.json");
+var server = require("../../server.js");
 
 var opts = {
     method: "GET",
-    url: "/issues/jrans"
+    url: "/issues/" + config.githubUsername
 };
 
-server.inject(opts, function (res) {
 
-    test("testing getting issues", function (t) {
-        var issues = res.payload;
+test("testing getting issues", function (t) {
+    server.inject(opts, function (res) {
+        var issues = JSON.parse(res.payload);
 
         t.equal(res.statusCode, 200, "status code is OK");
-        t.equal(issues.length > 0, true, "payload contains one or more issues");
+        t.ok(issues.length > 0, "payload contains one or more issues");
         t.end();
     });
+});
 
-    test("issues contain correct information", function (t) {
-        var issue = res.payload[0];
+test("issues contain correct information", function (t) {
+    server.inject(opts, function (res) {
+        var issues = JSON.parse(res.payload);
 
-        t.equal(issue.url.hasOwnProperty("url"), true, "contains the url to the issue");
-        t.equal(issue.url.hasOwnProperty("labels"), true, "contains labels prop");
+        t.ok(issues[0].hasOwnProperty("url"), "contains the url to the issue");
+        t.ok(issues[0].hasOwnProperty("labels"), "contains labels prop");
         t.end();
     });
 });
