+++ test/handlers/hello.test.js
@@ -1,5 +1,5 @@
 var test   = require('tape');
-var server = require('../server'); // test server which in turn loads our module
+var server = require('../../server.js'); // test server which in turn loads our module
 
 test("Hello World Test!", function(t) {
   var options = {
@@ -8,9 +8,9 @@ test("Hello World Test!", function(t) {
   };
   // server.inject lets us similate an http request
   server.inject(options, function(response) {
+
     t.equal(response.statusCode, 200, "Everything Works as expected!");
     t.equal(response.result, '<hello>Hello Benji!</hello>', "Response is Hello Benji!")
-    server.stop()
     t.end();
   });
 });
